+++ src/main/java/com/backendoori/ootw/weather/util/deserializer/ForecastResultHeaderDeserializer.java
@@ -1,7 +1,6 @@
 package com.backendoori.ootw.weather.util.deserializer;
 
 import java.io.IOException;
-import com.backendoori.ootw.weather.dto.forecast.ForecastResultHeader;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
@@ -13,8 +12,7 @@ public class ForecastResultHeaderDeserializer extends JsonDeserializer<ForecastR
 
     @Override
     public ForecastResultHeader deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
-        JsonNode node = p.getCodec()
-            .readTree(p);
+        JsonNode node = p.getCodec().readTree(p);
         JsonNode headerNode = node.findValue("response")
             .get("header");
