+++ src/main/java/com/backendoori/ootw/weather/util/deserializer/ForecastSuccessResultBody.java
@@ -1,7 +1,6 @@
-package com.backendoori.ootw.weather.dto.forecast;
+package com.backendoori.ootw.weather.util.deserializer;
 
 import java.util.List;
-import com.backendoori.ootw.weather.util.deserializer.ForecastSuccessResultBodyDeserializer;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 @JsonDeserialize(using = ForecastSuccessResultBodyDeserializer.class)
